add wrapper.c rakiyama.h
