+++ rakiyama/run.sh
@@ -1,3 +1,3 @@
 make -C ../libft
-gcc -I ../libft minipipe.c ../libft/libft.a
+gcc -I ../includes/ minipipe.c ../srcs/wrapper2.c ../libft/libft.a
 ./a.out
