+++ srcs/wrapper2.c
@@ -1,0 +1,78 @@
+#include "rakiyama.h"
+
+char	*ft_xstrjoin(char *str1, char *str2)
+{
+	char	*joined_str;
+
+	joined_str = ft_strjoin(str1, str2);
+	if (!joined_str)
+	{
+		perror("malloc");
+		exit(EXIT_FAILURE);
+	}
+	return (joined_str);
+}
+
+char	**ft_xsplit(char *src_str, char cut_char)
+{
+	char	**splited_str;
+
+	splited_str = ft_split(src_str, cut_char); 
+	if (!splited_str)
+	{
+		perror("malloc");
+		exit(EXIT_FAILURE);
+	}
+	return (splited_str);
+}
+
+void	xclose(int fd)
+{
+	if (close(fd) == -1)
+	{
+		ft_putstr_fd("close : ", STDERR_FILENO);
+		ft_putendl_fd(strerror(errno), STDERR_FILENO);
+		exit(EXIT_FAILURE);
+	}
+}
+
+void	xdup2(int oldfd, int newfd)
+{
+	if (dup2(oldfd, newfd) == -1)
+	{
+		ft_putstr_fd("dup2 : ", STDERR_FILENO);
+		ft_putendl_fd(strerror(errno), STDERR_FILENO);
+		exit(EXIT_FAILURE);
+	}
+}
+
+void	xwaitpid(pid_t pid, int *wstatus, int options)
+{
+	if (waitpid(pid, wstatus, options) == -1)
+	{
+		perror("waitpid");
+		exit(EXIT_FAILURE);
+	}
+}
+
+void	xpipe(int *pipefd)
+{
+	if (pipe(pipefd) == -1)
+	{
+		perror("pipe");
+		exit(EXIT_FAILURE);
+	}
+}
+
+pid_t	xfork(void)
+{
+	int	pid;
+
+	pid = fork();
+	if (pid == -1)
+	{
+		perror("fork");
+		exit(EXIT_FAILURE);
+	}
+	return (pid);
+}
